move IndexEdge/SnapshotEdge/Vertex Serializable, Deserializable into storage package from storage.hbase.
