+++ s2core/src/main/scala/com/kakao/s2graph/core/storage/Storage.scala
@@ -1,13 +1,11 @@
 package com.kakao.s2graph.core.storage
 
-import java.util.Base64
 
 
 import com.kakao.s2graph.core.ExceptionHandler.{Key, Val, KafkaMessage}
 import com.kakao.s2graph.core.GraphExceptions.FetchTimeoutException
 import com.kakao.s2graph.core._
 import com.kakao.s2graph.core.mysqls._
-import com.kakao.s2graph.core.storage.hbase._
 import com.kakao.s2graph.core.types._
 import com.kakao.s2graph.core.utils.{Extensions, logger}
 import com.typesafe.config.Config
@@ -929,7 +927,7 @@ abstract class Storage[R](val config: Config)(implicit ec: ExecutionContext) {
   protected def toRequestEdge(queryRequest: QueryRequest): Edge = {
     val srcVertex = queryRequest.vertex
     //    val tgtVertexOpt = queryRequest.tgtVertexOpt
-    val edgeCf = HSerializable.edgeCf
+    val edgeCf = Serializable.edgeCf
     val queryParam = queryRequest.queryParam
     val tgtVertexIdOpt = queryParam.tgtVertexInnerIdOpt
     val label = queryParam.label
