+++ s2core/src/main/scala/com/kakao/s2graph/core/storage/StorageDeserializable.scala
@@ -1,7 +1,7 @@
 package com.kakao.s2graph.core.storage
 
 import com.kakao.s2graph.core.utils.logger
-import com.kakao.s2graph.core.{Edge, IndexEdge, QueryParam}
+import com.kakao.s2graph.core.QueryParam
 import com.kakao.s2graph.core.types.{HBaseType, InnerVal, InnerValLike, InnerValLikeWithTs}
 import org.apache.hadoop.hbase.util.Bytes
